docker: by default, do not build when refreshing the rocket container

+++ docker/build.sh
@@ -1,5 +1,15 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 set -eu;
+
+set +u;
+
+if [[ ! -v "$nobuild" ]]
+then
+    nobuild='';
+fi;
+
+set -u;
+
 pwd="$(
     readlink -f "$(
         dirname "$(
@@ -7,9 +17,14 @@ pwd="$(
         )";
     )";
 )"'/..';
+
+(
 cd "$pwd";
 
-docker run --rm -it -v "$(pwd)":/home/rust/src ekidd/rust-musl-builder:nightly cargo build --release;
+if [[ -z "$nobuild" ]];
+then
+    docker run --rm -it -v "$(pwd)":/home/rust/src ekidd/rust-musl-builder:nightly cargo build --release;
+fi;
 
 platform='x86_64-unknown-linux-musl';
 out_name='target/'"$platform"'/release/'"$(
@@ -23,3 +38,4 @@ cp -rf templates target;
 cp -f Rocket.toml target;
 
 echo 'Build complete.';
+);
